// MDQ_API/convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * Generated by convex@1.13.2.
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";
import type * as MDQ from "../MDQ.js";
import type * as MDQActions from "../MDQActions.js";
import type * as MDQBids from "../MDQBids.js";
import type * as MDQBidsActions from "../MDQBidsActions.js";
import type * as MDQBidsMutations from "../MDQBidsMutations.js";
import type * as MDQBidsQueries from "../MDQBidsQueries.js";
import type * as MDQMutations from "../MDQMutations.js";
import type * as MDQQueries from "../MDQQueries.js";
import type * as api_ from "../api.js";
import type * as http from "../http.js";

/**
 * A utility for referencing Convex functions in your app's API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
declare const fullApi: ApiFromModules<{
  MDQ: typeof MDQ;
  MDQActions: typeof MDQActions;
  MDQBids: typeof MDQBids;
  MDQBidsActions: typeof MDQBidsActions;
  MDQBidsMutations: typeof MDQBidsMutations;
  MDQBidsQueries: typeof MDQBidsQueries;
  MDQMutations: typeof MDQMutations;
  MDQQueries: typeof MDQQueries;
  api: typeof api_;
  http: typeof http;
}>;
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;